tratamento de exceptions

+++ estudo-spring-java/src/main/java/com/example/estudo_spring_java/service/ProdutoService.java
@@ -1,10 +1,10 @@
 package com.example.estudo_spring_java.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.example.estudo_spring_java.exceptions.RecursoNaoEncontradoException;
 import com.example.estudo_spring_java.model.Produto;
 import com.example.estudo_spring_java.repository.ProdutoRespository;
 
@@ -21,8 +21,9 @@ public class ProdutoService {
         return produtoRespository.findAll();
     }
 
-    public Optional<Produto> buscarProdutoId(Long id) {
-        return produtoRespository.findById(id);
+    public Produto buscarProdutoId(Long id) {
+        return produtoRespository.findById(id)
+        .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID " +id+" não encontrado"));
     }
 
     public Produto SalvarProduto(Produto produto) {
@@ -30,6 +31,9 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id) {
+        if (!produtoRespository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Produto com ID " +id+ " não encontrado.");
+        }
         produtoRespository.deleteById(id);
     }
 }
